+++ frontend/src/components/ContextMenu.vue
@@ -23,7 +23,12 @@
           <i v-if="!showCreate" class="material-icons">add</i>
           <i v-if="showCreate" class="material-icons">arrow_back</i>
         </div>
-        <div v-if="selectedCount > 0" @mouseleave="hideTooltip" @mouseenter="showTooltip($event, $t('buttons.selectedCount'))" class="button selected-count-header">
+        <div
+          v-if="selectedCount > 0"
+          @mouseleave="hideTooltip"
+          @mouseenter="showTooltip($event, $t('buttons.selectedCount'))"
+          class="button selected-count-header"
+        >
           <span>{{ selectedCount }}</span>
         </div>
       </div>